separate out execute_operation
