--- quadratic-core/src/controller/execution/tests/mod.rs
@@ -1,5 +1,0 @@
-mod cell_value;
-mod multiplayer;
-mod offsets;
-mod sheets;
-mod spills;
