+++ quadratic-core/src/controller/thumbnail.rs
@@ -5,11 +5,11 @@ use super::GridController;
 impl GridController {
     /// whether the thumbnail needs to be updated for this Pos
     pub fn thumbnail_dirty_sheet_pos(&self, sheet_pos: SheetPos) -> bool {
-        self.thumbnail_dirty_sheet_rect(sheet_pos.into())
+        self.thumbnail_dirty_sheet_rect(&sheet_pos.into())
     }
 
     /// whether the thumbnail needs to be updated for this rectangle
-    pub fn thumbnail_dirty_sheet_rect(&self, sheet_rect: SheetRect) -> bool {
+    pub fn thumbnail_dirty_sheet_rect(&self, sheet_rect: &SheetRect) -> bool {
         if sheet_rect.sheet_id != self.grid().first_sheet_id() {
             return false;
         }
@@ -54,12 +54,12 @@ mod test {
     fn test_thumbnail_dirty_rect() {
         let gc = GridController::new();
         let sheet_id = gc.sheet_ids()[0];
-        assert!(gc.thumbnail_dirty_sheet_rect(SheetRect {
+        assert!(gc.thumbnail_dirty_sheet_rect(&SheetRect {
             min: Pos { x: 0, y: 0 },
             max: Pos { x: 1, y: 1 },
             sheet_id,
         }));
-        assert!(!gc.thumbnail_dirty_sheet_rect(SheetRect {
+        assert!(!gc.thumbnail_dirty_sheet_rect(&SheetRect {
             min: Pos {
                 x: (THUMBNAIL_WIDTH as i64) + 1i64,
                 y: 0
@@ -70,7 +70,7 @@ mod test {
             },
             sheet_id,
         }));
-        assert!(!gc.thumbnail_dirty_sheet_rect(SheetRect {
+        assert!(!gc.thumbnail_dirty_sheet_rect(&SheetRect {
             min: Pos {
                 x: 0,
                 y: (THUMBNAIL_HEIGHT as i64) + 1i64,
@@ -82,7 +82,7 @@ mod test {
             sheet_id,
         }));
         assert!(!gc.thumbnail_dirty_sheet_rect(
-            SheetPos {
+            &SheetPos {
                 x: THUMBNAIL_WIDTH as i64,
                 y: THUMBNAIL_HEIGHT as i64,
                 sheet_id
