+++ quadratic-core/src/grid/sheet/cell_array.rs
@@ -24,7 +24,9 @@ impl Sheet {
 
     /// In a given rect, collect all cell values into an array.
     ///
-    /// TODO(ddimaria): is this necessary as it's more performant to just pluck the data from the sheet direclty
+    /// TODO(ddimaria): is this necessary as it's more performant to just pluck the data from the sheet directly
+    /// davidfig: regrettably, the Array::new_row_major requires the ordering to be row-based and not column based.
+    /// we would need to rework how Array works for this to be more performant.
     pub fn cell_values_in_rect(&self, &selection: &Rect) -> Result<Array> {
         let values = selection
             .y_range()
